Add field level encryption configuration to Cloudfront
This adds field level encryption to the Cloudfront distribution.
Currently it applies globally to any `POST` request with `Content-Type:
application/x-www-form-urlencoded`. The form fields `cardNo`, `cvc`,
`expiryMonth`, `expiryYear` and `addressPostcode` will be encrypted
using the provisioned public key.

Unfortunately the Terraform AWS provider doesn't currently support field
level encryption configuration so I've had to hack it a bit with an
external data resource and a Ruby script that provisions the FLE profile
and config for the given public key.