PP-6127 Add secret service for publicapi
Implement version 1 of provisioning secrets in paas. It is likely that
this will be replaced when PaaS provide a more robust means for secrets
management.

This adds a credentials module which will gather secrets from
pay-low-pass. The credentials are set on a user provided service which
will be bound to the app.

To get it started this commit deals with publicapi but the same approach
and be used for the other apps.

with @rjbaker