PP-6561 Fix cardid cloning of data submodule
cardid-data submodule's url is defined as ssh yet concourse
is using oauth. Furthermore we need to avoid the password
prompt from the git cli. Therefore rewrite the url for https
and add the token. The risk of setting the token in the url is
mitigated since these files are not committed, the container is ephemeral
and anyone with access to read the files could read the token from
environment variable. Furthermore we redact the token from the
files after the update.